+++ src/main/java/gov/ca/cwds/cals/service/dto/rfa/OtherAdultDTO.java
@@ -49,6 +49,9 @@ public class OtherAdultDTO extends RFAExternalEntityDTO {
   @Valid
   private List<RelationshipToApplicantDTO> relationshipToApplicants = new ArrayList<>();
 
+  @ApiModelProperty(hidden = true)
+  private List<RFA1bFormDTO> rfa1bForms;
+
   public NamePrefixType getNamePrefix() {
     return namePrefix;
   }
@@ -106,4 +109,13 @@ public class OtherAdultDTO extends RFAExternalEntityDTO {
   public void setNameSuffix(NameSuffixType nameSuffix) {
     this.nameSuffix = nameSuffix;
   }
+
+  public List<RFA1bFormDTO> getRfa1bForms() {
+    return rfa1bForms;
+  }
+
+  public void setRfa1bForms(List<RFA1bFormDTO> rfa1bForms) {
+    this.rfa1bForms = rfa1bForms;
+  }
+  
 }
